refactor to aws-sdk-go-v2, add explicit region config
The SDK was updated to the V2 version in order to better support
handling the configuration of the SDK.

Added a new FS_REGION parameter which will set the AWS_REGION for this
program specifically, this is for applications which have different
regions for secrets and other uses.

Improved logging around some errors.